Use combinations to test eager/graph mode and TF v1/v2 for filter tests
PiperOrigin-RevId: 263814675
